pyproject.toml
[tool.poetry]
name = "ml-templete"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
networkx = "^3.3"
jupyter = "^1.0.0"
notebook = "^7.2.1"
umap-learn = "^0.5.6"
scikit-learn = "^1.5.1"
ipykernel = "^6.29.5"
matplotlib = "^3.9.2"
torch-geometric = "^2.6.0"
torch = "^2.4.1"
torchvision = "^0.19.1"
python-dotenv = "^1.0.1"
openai = "^1.57.4"
pymupdf4llm = "^0.0.17"
notion-client = "^2.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
